aiml-service/requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.30.1
sentence-transformers==3.0.1
torch==2.3.1
torchvision==0.18.1
scikit-learn==1.5.1
pillow==10.4.0
numpy==2.0.0
